requirements.txt
blinker==1.7.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
dlib==19.24.2
dropbox==11.36.2
face-recognition==1.3.0
firebase-admin==6.2.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
Flask==3.0.0
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth-httplib2==0.1.1
google-auth==2.24.0
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio-status==1.59.3
grpcio==1.59.3
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msgpack==1.0.7
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.8.1.78
passlib==1.7.4
Pillow==10.1.0
ply==3.11
proto-plus==1.22.3
protobuf==4.25.1
psycopg2-binary==2.9.9
pyasn1-modules==0.3.0
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.0
requests-oauthlib==1.3.1
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.23
stone==3.3.1
tensorflow==2.15.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
Werkzeug==3.0.1